Move ProgressTimeout class to utils
